Use tri-state CMake variable for SIMD options
Define a tri-state option variable for SIMD support in CMake:
- Disable: disable support for the SIMD feature entirely (compiled out).
- RuntimeCheckEnabled: enable support for the SIMD feature, with runtime
check for availability using CPUID. This is the default for most options
and the safest.
- CompileTimeAssumed: enable support for the SIMD feature, and assume
compile-time support, directly using the CPU instructions. This must
only be used with features that are assumed to always be present on the
target device. This is the default for SSE and SSE2, which are old
enough to be supported by any modern CPU.